fix position columns' data type to be double

+++ database/migrations/2025_04_29_154246_create_columns_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
       $table->id();
       $table->foreignIdFor(Board::class)->constrained()->cascadeOnDelete();
       $table->string("title");
-      $table->unsignedInteger('position');
+      $table->double('position');
       $table->timestamps();
     });
   }
